Moved a function into dirshared.cpp
